requirements.txt
-e .
aws-cdk.aws-lambda
aws-cdk.aws-sns
aws-cdk.aws-dynamodb
aws-cdk.aws-lambda-python
aws-cdk.aws-lambda-event-sources
diagrams
pynamodb
boto3